pyproject.toml
[project]
name = "mtglimited"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.7",
    "mypy>=1.13.0",
    "numpy>=2.1.2",
    "pylint>=3.3.1",
    "pymongo>=4.10.1",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "types-requests>=2.32.0.20241016",
]
